change likes to mark

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -16,7 +16,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 @ToString
-@EqualsAndHashCode(of = "id")
+@EqualsAndHashCode(of = "id") // Объект определяется только по полю id
 public class Film {
 
     private long id;
@@ -32,7 +32,7 @@ public class Film {
     private LocalDate releaseDate;
     @Min(1)
     private int duration;
-    private Integer rate;
+    private Double rate;
     private List<Director> directors = new ArrayList<>();
 
 }
